# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gyroscope real-time monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /motor/command:
    post:
      operationId: sendMotorCommand
      tags: [motor]
      summary: Send a command to the motor controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotorCommandInput"
      responses:
        "200":
          description: Command acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MotorCommandResponse"
...
  /motor/setpoint:
    post:
      operationId: setMotorSetpoint
      tags: [motor]
      summary: Set manual RPM target or switch back to auto mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualRpmInput"
      responses:
        "200":
          description: Setpoint applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualRpmResponse"
        "400":
          description: Invalid input
  /motor/pid:
    post:
      operationId: setPidEnabled
      tags: [motor]
      summary: Toggle PID auto-stabilization on/off without changing the current throttle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PidToggleInput"
      responses:
        "200":
          description: PID mode toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PidToggleResponse"
        "400":
          description: Invalid input
...
components:
  schemas:
...
    MotorCommandInput:
      type: object
      required: [command]
      properties:
        command:
          type: string
          enum: [on, off, emergency_stop]
    MotorCommandResponse:
      type: object
      required: [ok, motorState, message]
      properties:
        ok:
          type: boolean
        motorState:
          type: string
          enum: [on, off, emergency_stopped]
        message:
          type: string
...
    ManualRpmInput:
      type: object
      required: [targetRpm, mode]
      properties:
        targetRpm:
          type: integer
          minimum: 0
          maximum: 12000
          description: Target flywheel RPM for manual control (0 = stop)
        mode:
          type: string
          enum: [auto, manual]
          description: auto = PID loop controls RPM; manual = direct RPM setpoint
    ManualRpmResponse:
      type: object
      required: [ok, targetRpm, mode, message]
      properties:
        ok:
          type: boolean
        targetRpm:
          type: integer
        mode:
          type: string
          enum: [auto, manual]
        message:
          type: string
    PidToggleInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
          description: true = enable PID auto-stabilization; false = disable (holds current throttle)
    PidToggleResponse:
      type: object
      required: [ok, mode, message]
      properties:
        ok:
          type: boolean
        mode:
          type: string
          enum: [auto, manual]
        message:
          type: string
...